Sign up form project
Finished the major features and look of the website. Current things to
fix are the logo since I want to make it white to contrast with the
background, and a proper text for the logo. Next changes are handling
invalid entries, specifically  making specific error messages and
changing the input color, making certain fields required and give
a visual indication for it.