Added styling for form validation
Only issues left are: cannot check if password equals confirm
password since that requires JS and I'm not using it for this particular
project. Also, the phone number field doesn't quite work as intended
since non-numeric inputs are considered valid. Also, the input fields
move when the error message appears. I am not quite sure how to fix that.
I'll leave it since this is just a practice project.