Developed and implemented page structure outline
It is a conceptual idea of what the page should look like.
The bottom part (where the books go) will be able to be changed
betwen a list mode and card mode. User will be able to add a new book
via a modal window that will appear when the player clicks the add
books button